requirements.txt
av==6.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1613413867554/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1610093487176/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1620867593185/work
colorama==0.4.6
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1616851478755/work
cycler==0.11.0
decorator @ file:///tmp/build/80754af9/decorator_1620904181799/work
decord==0.6.0
docopt==0.6.2
dominate==2.7.0
einops==0.6.1
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1589994591731/work
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1617858305498/work
humanize @ file:///home/conda/feedstock_root/build_artifacts/humanize_1619801300956/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1619012194154/work
ipdb @ file:///home/conda/feedstock_root/build_artifacts/ipdb_1618802261930/work
ipython @ file:///tmp/build/80754af9/ipython_1617118429768/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi==0.17.0
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
jsonpickle==1.5.2
kiwisolver==1.4.4
matplotlib==3.5.3
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1618853974840/work
mkl-service==2.3.0
msgpack @ file:///tmp/build/80754af9/msgpack-python_1612287175283/work
munch==2.5.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1618497199865/work
olefile==0.46
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1612459636436/work
pandas==1.1.4
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
prefetch-generator==1.0.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
psutil @ file:///tmp/build/80754af9/psutil_1612298016854/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py-cpuinfo==9.0.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1610291444829/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==6.0
regex @ file:///home/conda/feedstock_root/build_artifacts/regex_1617644420360/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
sacred==0.8.2
sacremoses @ file:///home/conda/feedstock_root/build_artifacts/sacremoses_1588857588528/work
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1614445567925/work
scipy @ file:///tmp/build/80754af9/scipy_1618852618548/work
six @ file:///tmp/build/80754af9/six_1605205313296/work
smmap==5.0.0
threadpoolctl @ file:///tmp/tmp79xdzxkt/threadpoolctl-2.1.0-py3-none-any.whl
timm==0.6.13
tokenizers @ file:///home/conda/feedstock_root/build_artifacts/tokenizers_1615295871129/work
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1617680194207/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1620845405335/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1615828766818/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
wrapt==1.15.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1614945704755/work
